# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Product Rental API",
        "description": "API for managing product rentals with regional pricing",
        "contact": {
            "name": "API Support",
            "email": "contact@example.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/rentals": {
            "get": {
                "tags": [
                    "Rentals"
                ],
                "summary": "Get list of rentals",
                "description": "Returns paginated list of rental transactions with optional filtering",
                "operationId": "3e1e61909e3d63891b376d33996755e8",
                "parameters": [
                    {
                        "name": "product_id",
                        "in": "query",
                        "description": "Filter by product ID",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "region_id",
                        "in": "query",
                        "description": "Filter by region ID",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filter by status (confirmed, cancelled, completed)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "confirmed",
                                "cancelled",
                                "completed"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Rentals"
                ],
                "summary": "Create new rental",
                "description": "Creates a new rental transaction with customer information",
                "operationId": "19e770ee112b4e0029f66c45f23f53e7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "product_id",
                                    "region_id",
                                    "rental_period_id",
                                    "customer_name",
                                    "customer_email",
                                    "customer_address",
                                    "start_date"
                                ],
                                "properties": {
                                    "product_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "region_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "rental_period_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "customer_name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "customer_email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "john@example.com"
                                    },
                                    "customer_address": {
                                        "type": "string",
                                        "example": "123 Main St, City, Country"
                                    },
                                    "start_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-04-15"
                                    },
                                    "notes": {
                                        "type": "string",
                                        "example": "Please deliver in the morning",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Rental created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Product rental created successfully"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error or product not available for rental"
                    }
                }
            }
        },
        "/api/v1/rentals/{rental}": {
            "get": {
                "tags": [
                    "Rentals"
                ],
                "summary": "Get rental details",
                "description": "Returns detailed information about a specific rental transaction",
                "operationId": "6b9ad001af9d384e98f4ab4eef48a711",
                "parameters": [
                    {
                        "name": "rental",
                        "in": "path",
                        "description": "Rental ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Rental not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Rentals"
                ],
                "summary": "Update rental",
                "description": "Updates an existing rental transaction",
                "operationId": "d247aaed186c5f39f0317f4da24dbdf1",
                "parameters": [
                    {
                        "name": "rental",
                        "in": "path",
                        "description": "Rental ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "customer_name": {
                                        "type": "string",
                                        "example": "John Smith"
                                    },
                                    "customer_email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "customer_address": {
                                        "type": "string"
                                    },
                                    "notes": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "confirmed",
                                            "cancelled",
                                            "completed"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Rental updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Rental updated successfully"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Rental not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        }
    },
...
}